Active MQ/Fuse Stomp listener example. Originally develeped by Abe Voelker.
This example Connnect to MQ server and listens for incomming packets on durable topic "cwh.data".
It identifies as "Client-id=cwh, Subscription Name=cwh" to subscribe durable topic.

If there is incloming packet of connected port reads its header and body in MessageHandler procedure.
If incoming packet is File, divided on parts and creates Files on local disk and concatenate parts.

Works with FileToMq.p procedure.